+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 4.2.0
 * Add `captureId` to `LocalPaymentDetails`
 * Add `debugId` to `LocalPaymentDetails`
 * Add `refundId` to `LocalPaymentDetails`